Rename file_utils to connector_utils
