+++ packages/terafoundation/lib/connector_utils.js
